Criação do ErrorResponse. Implementação do serviço de sessão do usuário e validação da mesma nos outros serviços. Melhoria nos serviços existentes.

+++ Code-DD-UFG/Back-End/sdd-ufg/src/main/java/br/ufg/inf/sdd_ufg/resource/KnowledgeLevelResource.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Map;
 
 import javax.inject.Inject;
+import javax.persistence.EntityExistsException;
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -17,27 +18,41 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
+import br.ufg.inf.sdd_ufg.dao.KnowledgeGroupDao;
 import br.ufg.inf.sdd_ufg.dao.KnowledgeLevelDao;
+import br.ufg.inf.sdd_ufg.dao.TeacherDao;
+import br.ufg.inf.sdd_ufg.model.KnowledgeGroup;
 import br.ufg.inf.sdd_ufg.model.KnowledgeLevel;
+import br.ufg.inf.sdd_ufg.model.Teacher;
 import br.ufg.inf.sdd_ufg.resource.utils.ResultSetResponse;
 
-@Path("/knowledgeLevels")
+@Path("/knowledge_levels")
 @Produces(MediaType.APPLICATION_JSON)
 public class KnowledgeLevelResource extends AbstractResource {
 
 	private final KnowledgeLevelDao knowledgeLevelDao;
+	private final TeacherDao teacherDao;
+	private final KnowledgeGroupDao knowledgeGroupDao;
 
     @Inject
-    public KnowledgeLevelResource(final KnowledgeLevelDao knowledgeLevelDao) {
+    public KnowledgeLevelResource(final KnowledgeLevelDao knowledgeLevelDao,
+    		final TeacherDao teacherDao, 
+    		final KnowledgeGroupDao knowledgeGroupDao) {
         this.knowledgeLevelDao = knowledgeLevelDao;
+        this.teacherDao = teacherDao;
+        this.knowledgeGroupDao = knowledgeGroupDao;
     }
 	
     @GET
 	@Path("/{id}")
 	public Response retrieveKnowledgeLevelById(@PathParam("id") Long id, @Context final HttpServletRequest request) {
-		KnowledgeLevel knowledgeLevel = knowledgeLevelDao.findById(id, 1);
+    	if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
+		}
+    	
+    	KnowledgeLevel knowledgeLevel = knowledgeLevelDao.findById(id, 1);
 		if (knowledgeLevel == null) {
-			return Response.status(Response.Status.NOT_FOUND).build();
+			return getResourceNotFoundResponse();
 		}
 		return Response.ok(knowledgeLevel)
 				.build();
@@ -45,6 +60,10 @@ public class KnowledgeLevelResource extends AbstractResource {
     
 	@GET
     public Response retrieveAllKnowledgeLevels(@QueryParam("page") Integer page, @Context final HttpServletRequest request) {
+		if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
+		}
+		
 		List<KnowledgeLevel> knowledgeLevels = knowledgeLevelDao.findAll(0);
 		if (knowledgeLevels == null || knowledgeLevels.size() == 0) {
 			return Response.status(Response.Status.NOT_FOUND)
@@ -61,13 +80,18 @@ public class KnowledgeLevelResource extends AbstractResource {
 	
 	@POST
 	public Response insertKnowledgeLevel(@Context final HttpServletRequest request) {
+		if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
+		}
+		
 		KnowledgeLevel knowledgeLevel;
 		try {
 			knowledgeLevel = retrieveKnowledgeLevelFromJson(request);
 			knowledgeLevelDao.insert(knowledgeLevel);
+		} catch (EntityExistsException eee) {
+			return getInsertErrorResponse();
 		} catch (Exception e) {
-			return Response.status(Response.Status.BAD_REQUEST)
-					.build();
+			return getBadRequestResponse();
 		}
 		return Response.ok(knowledgeLevel)
 				.build();
@@ -76,17 +100,21 @@ public class KnowledgeLevelResource extends AbstractResource {
 	@PUT
 	@Path("/{id}")
 	public Response updateKnowledgeLevel(@PathParam("id") Long id, @Context final HttpServletRequest request) {
+		if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
+		}
+		
 		KnowledgeLevel knowledgeLevel;
 		try {
 			knowledgeLevel = retrieveKnowledgeLevelFromJson(request);
 			knowledgeLevel.setId(id);
 			knowledgeLevelDao.update(knowledgeLevel);
+		} catch (EntityExistsException eee) {
+			return getInsertErrorResponse();
 		} catch (IllegalArgumentException iae) {
-			return Response.status(Response.Status.NOT_FOUND)
-					.build();
+			return getResourceNotFoundResponse();
 		} catch (Exception e) {
-			return Response.status(Response.Status.BAD_REQUEST)
-					.build();
+			return getBadRequestResponse();
 		}
 		return Response.ok(knowledgeLevel)
 				.build();
@@ -96,8 +124,12 @@ public class KnowledgeLevelResource extends AbstractResource {
 		Map<String, Object> content = getJSONContent(request);
 		
 		KnowledgeLevel knowledgeLevel = new KnowledgeLevel();
-		//knowledgeLevel.setTeacher(content.get("name").toString());
-		//knowledgeLevel.setKnowledgeGroup(content.get("cpf").toString());
+		
+		Teacher teacher = teacherDao.findById(new Long(content.get("teacher_id").toString()), 0);
+		knowledgeLevel.setTeacher(teacher);
+		
+		KnowledgeGroup knowledgeGroup = knowledgeGroupDao.findById(new Long(content.get("knowledge_id").toString()), 0);
+		knowledgeLevel.setKnowledgeGroup(knowledgeGroup);
 		
 		return knowledgeLevel;
 	}
@@ -105,11 +137,14 @@ public class KnowledgeLevelResource extends AbstractResource {
 	@DELETE
 	@Path("/{id}")
 	public Response deleteKnowledgeLevel(@PathParam("id") Long id, @Context final HttpServletRequest request) {
+		if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
+		}
+		
 		try {
 			knowledgeLevelDao.delete(id);
 		} catch (IllegalArgumentException iae) {
-			return Response.status(Response.Status.NOT_FOUND)
-					.build();
+			return getResourceNotFoundResponse();
 		}
 		return Response.status(Response.Status.NO_CONTENT)
 				.build();
